Moved TabController to its own file, added TabController onChanged() listeners
